pyproject.toml
[tool.poetry]
name = "swamp-discord-bot"
version = "0.1.0"
description = "A Discord bot for managing a friend server"
authors = ["Forest <adsf@a.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.11.11"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
pylint = "^3.3.2"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
